pyproject.toml
[project]
name = "threads-ai-agent"
version = "0.1.0"
description = "AI agent for creating Threads posts using web scraping"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "boto3>=1.34.0",
    "click>=8.1.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "playwright>=1.40.0",
    "parsel>=1.8.0",
    "jmespath>=1.0.0",
    "nested-lookup>=0.2.25",
    "nest-asyncio>=1.6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
threads-agent = "threads_ai_agent.cli:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
